requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-filter==21.1
drf-yasg==1.20.0
djangorestframework-simplejwt==5.1.0
django-countries==7.3.2
